Fix error where TwainScanner cleans out folders
This fixes an issue discovered in Financial where the scanner was taking out scanned files that are within the "OCR" folder which is considered as a "custom folder" to delete everything out of. This Will probably require a follow up fix to allow for proper cleaning of temp items from folders.